+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   includes a `CliUtilDialogSupport` implementation for use in simple CLIs.
 - feat: **Breaking Change** - Removed `stdout` and `stdin` parameters to
   `getSkills` and replaced them with a required `Logger logger`.
+- feat: Check packages for security advisories on install.
 - chore: Move cache dir to `.dart_tool/skills` from `.dart_skills`.
 - feat: Allow the user to select an IDE if none is detected.